Fix parent hash empty case in parents validation
Note the only time that genesis validation happens is
at the beginning of the chain. Hence, this would
never really have appeared as a bug: it just makes
no logical sense to label a block with an empty parent
hash as a last finalized block that isn't genesis -
as that would have a parent.

I've also added some logging to debug the invalid parents
warning we are getting on testnet.