Change number of consumers that handle incoming queue of blocks to 1
An interesting observation was fact that some callback were held by
MultiparentCasperImpl semaphore.

This itself is not an issue. But when you look how TransportLayer is
propagating received blocks to casper. Each incoming message (block)
is stored on the file system. There is a server queue to which an
information is added that given block was stored. Then there is a
number of consumers that consume elements from the queue, read
messages from the filesystem and push calculation to casper.

Detail in https://rchain.atlassian.net/browse/RCHAIN-2876?focusedCommentId=13463&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-13463