Introduce lazy serializedSizeM
Issue (what was):

Memo class is implements an memoization concept. It has a thunk that will be “nulled” (for GC to collect it).
The problem is, that memory is freed only when calculcation is executed (get method is called and evaluated). Which is not always the case (in memory block storage, genesis block in memory). All proto generated objects have that memo value added.
This can eat up 10% of all heap!

Fix (what is):

We init the value when it's needed (and also evaluated thus freeing memory)