Use GCA instead of LFB in estimator
The issue for the invalid parents right now is that the last finalized block for each validator is different. And so it is possible that even though you know that the main chain will eventually settle on a certain block B, another validator might add a block that uses something before B as a parent.
So for the purposes of calculating GHOST, we don't need a last finalized block as much as a greatest common ancestor block of all validators. The GCA block could be either before or after the last finalized block on the main chain. The GCA block may surprisingly also not be in the main chain eventually.