Simplify rholang source loading
Previously, a custom sbt module would generate a Scala class for
each .rho file. This would lead to 'duplicate zip entry' errors when
building with interchangeably with sbt and IntelliJ IDEA, and even
without that cause the feedback loop for rholang tests to be ~60 s by
demanding casper/test:compile each time.

With this commit, a standard mechanism of classpath resources is used
for storing and accessing the .rho files, allowing for rapid changes
in .rho files and tests to be seen by IDEA immediately.

Lastly, this removes a bit of work and a module from the sbt build,
making it a bit faster.