Postpone Casper initialization until comm receives first connection
CasperPacketHandler is seriously the weirdest set of design choices
I've seen in the past few weeks. I'm planning to refactor hell out of
it in incoming days, but for now this commit is focusing on fixing bug
RCHAIN-3007.

After this change, when initializing node (which is not bootstrap),
casper code will not be triggered until comm layer grabs at least
first connection. In 99% of cases the flow of the program will be
following:

1. node boots up
2. node puts bootstrap into its kademlia table
3. node waits for connection to bootstrap node
4. once connection is established:
a) CasperPachketHandler protocol is run (init method)
b) casperloop is triggered

After this change, new node will start catching up with the network
even if nobody is proposing to it (or other nodes) at the given moment