Remove block buffer and instead use dependencyDag
It no longer makes sense to keep the block buffer and
dependencyDag as two separate data structures. Previously,
we needed the block buffer because it stored entire
blocks but now since it only stores hashes, we can extract
the same information from dependencyDag by adding a contains
method to it.