Bring back idea of seen block hashes through dependencyDag
This fixes a subtle bug in which the dependencyFree blocks
have not necessarily been processed and added to the
block store.

Unfortunately, the dependencyDag still needs to keep
information about which blocks actually came in and which
hashes are in the dependencyDag because they are dependencies
of other blocks but aren't yet seen. But instead of keeping
a separate data structure we can just use
`state.dependencyDag.parentToChildAdjacencyList.values` as
what was previously the block buffer (seen blocks) as
all seen blocks are inserted into the dependencyDag as
children.