Add scalafmtCheckAll to pr build (bors/branch build to come later)
This will fail the PR build if code isn't formatted properly.
This won't fail builds done by bors, so that it fulfills only
informational role for now. We might decide to enforce that for in bors
builds later, depending on the results.

Parallel execution is disabled for scalafmtCheckAll in CI.
See https://github.com/scalameta/scalafmt/issues/1399 for reasons why.