Make the assertions in CostAccountingSpec more robust
Previously we asserted what operation would tip it past the phloLimit.
With parallel execution though this can cause flaky tests. Say we run:

@0!(0) | @0!(0)

and the operations for each of the pars are: "send eval", "substitute".

If we execute with limit L = parsing cost + "send eval" cost,
then we can't assert which operation will go past the limit, as it could
be either "substitute" from the first executed par, or "send eval" from
the other.

Previously the tests were passing consistently, because all pars were
first charging twice for substitution. This is no longer the case, and
the first two operations a par charges for are no longer "do X twice".

If you've read this far, please let me know if I should've bothered
writing that. Thx!